Add optional userid to scripts for remote instances
EC2 instances running Ubuntu require signing on as 'ubuntu',
while instances running Amazon Linux require signing on as 'ec2-user'.

This revision adds an optional userid to `signon.sh`.  The script
saves that userid to be used by `transfer.sh` whenever a file is
to be sent to the current remote instance.